🧬 Migrate the utility functions about processing JavaScript attributes into another new module to manage it.
