+++ pymock_api/model/openapi/_js_handlers.py
@@ -1,0 +1,20 @@
+def convert_js_type(t: str) -> str:
+    if t == "string":
+        return "str"
+    elif t in ["integer", "number"]:
+        return "int"
+    elif t == "boolean":
+        return "bool"
+    elif t == "array":
+        return "list"
+    elif t == "file":
+        return "file"
+    else:
+        raise TypeError(f"Currently, it cannot parse JS type '{t}'.")
+
+
+# TODO: Should clean the parsing process
+def ensure_type_is_python_type(t: str) -> str:
+    if t in ["string", "integer", "number", "boolean", "array"]:
+        return convert_js_type(t)
+    return t
