+++ pymock_api/model/openapi/_parser.py
@@ -8,30 +8,9 @@ from ._base import (
     ensure_get_schema_parser_factory,
     get_openapi_version,
 )
+from ._js_handlers import convert_js_type, ensure_type_is_python_type
 from ._parser_factory import BaseOpenAPISchemaParserFactory
 from ._schema_parser import BaseOpenAPIPathSchemaParser, BaseOpenAPISchemaParser
-
-
-def convert_js_type(t: str) -> str:
-    if t == "string":
-        return "str"
-    elif t in ["integer", "number"]:
-        return "int"
-    elif t == "boolean":
-        return "bool"
-    elif t == "array":
-        return "list"
-    elif t == "file":
-        return "file"
-    else:
-        raise TypeError(f"Currently, it cannot parse JS type '{t}'.")
-
-
-# TODO: Should clean the parsing process
-def ensure_type_is_python_type(t: str) -> str:
-    if t in ["string", "integer", "number", "boolean", "array"]:
-        return convert_js_type(t)
-    return t
 
 
 class APIParser:
